# perf: Reduce allocations and printfs in G-function cache.