Previously, the G-function cache used repeated fmt.Sprintf("%v", x) and string concatenation to generate its cache key. This resulted in lots of unnecessary allocations, both in the process of Sprintf and in an O(n²) string building loop. This change allocates everything in one go and relies on the fact that we already know everything is strings.