@thetorpedodog 2022-02-17
…actually, upon further consideration, we don’t need to concatenate strings at all; we can just use a composite struct key. This further reduces the number of allocations we need to make!