@hsluoyz 2022-02-18
@thetorpedodog can you paste the benchmark diff here? https://casbin.org/docs/en/benchmark