@thetorpedodog 2022-02-18
Before:
```
goos: linux
goarch: amd64
pkg: github.com/casbin/casbin/v2
cpu: Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz
BenchmarkCachedRaw-4                                            76673332                15.70 ns/op            0 B/op          0 allocs/op
BenchmarkCachedBasicModel-4                                      8203942               148.0 ns/op           104 B/op          4 allocs/op
BenchmarkCachedRBACModel-4                                       8185062               146.4 ns/op           104 B/op          4 allocs/op
BenchmarkCachedRBACModelSmall-4                                  7929399               150.3 ns/op           104 B/op          4 allocs/op
BenchmarkCachedRBACModelMedium-4                                 7978339               150.0 ns/op           104 B/op          4 allocs/op
BenchmarkCachedRBACModelLarge-4                                  8778506               133.1 ns/op            96 B/op          3 allocs/op
BenchmarkCachedRBACModelWithResourceRoles-4                      8018532               151.4 ns/op           104 B/op          4 allocs/op
BenchmarkCachedRBACModelWithDomains-4                            7484630               160.2 ns/op           120 B/op          4 allocs/op
BenchmarkCachedABACModel-4                                        241495              5271 ns/op            2722 B/op         47 allocs/op
BenchmarkCachedKeyMatchModel-4                                   7015843               166.5 ns/op           152 B/op          4 allocs/op
BenchmarkCachedRBACModelWithDeny-4                               7834410               152.2 ns/op           104 B/op          4 allocs/op
BenchmarkCachedPriorityModel-4                                   8040786               150.7 ns/op           104 B/op          4 allocs/op
BenchmarkCachedRBACModelMediumParallel-4                        15450997                74.62 ns/op          105 B/op          4 allocs/op
BenchmarkHasPolicySmall-4                                        2353174               501.3 ns/op           182 B/op          8 allocs/op
BenchmarkHasPolicyMedium-4                                       2276661               538.1 ns/op           189 B/op          8 allocs/op
BenchmarkHasPolicyLarge-4                                        2174547               549.2 ns/op           197 B/op          9 allocs/op
BenchmarkAddPolicySmall-4                                        2249992               534.9 ns/op           184 B/op          8 allocs/op
BenchmarkAddPolicyMedium-4                                       1968679               592.2 ns/op           205 B/op          9 allocs/op
BenchmarkAddPolicyLarge-4                                        1000000              1056 ns/op             506 B/op         11 allocs/op
BenchmarkRemovePolicySmall-4                                     2341110               512.8 ns/op           198 B/op          9 allocs/op
BenchmarkRemovePolicyMedium-4                                    1966756               579.5 ns/op           209 B/op          9 allocs/op
BenchmarkRemovePolicyLarge-4                                     1000000              1064 ns/op             324 B/op         15 allocs/op
BenchmarkRaw-4                                                  76758178                15.68 ns/op            0 B/op          0 allocs/op
BenchmarkBasicModel-4                                             128106              9364 ns/op            4915 B/op         90 allocs/op
BenchmarkRBACModel-4                                               88168             13060 ns/op            6086 B/op        128 allocs/op
BenchmarkRBACModelSmall-4                                           9385            123020 ns/op           44905 B/op       1400 allocs/op
BenchmarkRBACModelMedium-4                                          1065           1121964 ns/op          455403 B/op      13126 allocs/op
BenchmarkRBACModelLarge-4                                            100          11545071 ns/op         4399639 B/op     130265 allocs/op
BenchmarkRBACModelWithResourceRoles-4                              91244             12739 ns/op            6801 B/op        118 allocs/op
BenchmarkRBACModelWithDomains-4                                    77934             15269 ns/op            8102 B/op        144 allocs/op
BenchmarkABACModel-4                                              241245              5003 ns/op            2713 B/op         46 allocs/op
BenchmarkKeyMatchModel-4                                           85480             14154 ns/op            8131 B/op        122 allocs/op
BenchmarkRBACModelWithDeny-4                                       81309             14779 ns/op            6543 B/op        148 allocs/op
BenchmarkPriorityModel-4                                          110157             10901 ns/op            5562 B/op        103 allocs/op
BenchmarkRBACModelWithDomainPatternLarge-4                         35865             32766 ns/op           23102 B/op        280 allocs/op
BenchmarkRoleManagerSmall-4                                        29073             42050 ns/op           13539 B/op        498 allocs/op
BenchmarkRoleManagerMedium-4                                        2758            431509 ns/op          141912 B/op       5742 allocs/op
BenchmarkRoleManagerLarge-4                                          264           4426323 ns/op         1509933 B/op      59742 allocs/op
BenchmarkBuildRoleLinksWithPatternLarge-4                              1        1162629378 ns/op        688769208 B/op  10085340 allocs/op
BenchmarkBuildRoleLinksWithDomainPatternLarge-4                      168           7204872 ns/op         3781946 B/op      89047 allocs/op
BenchmarkBuildRoleLinksWithPatternAndDomainPatternLarge-4              1        1175799088 ns/op        690574704 B/op  10120305 allocs/op
BenchmarkHasLinkWithPatternLarge-4                                 68646             17274 ns/op            4904 B/op        154 allocs/op
BenchmarkHasLinkWithDomainPatternLarge-4                         1000000              1197 ns/op             528 B/op         13 allocs/op
BenchmarkHasLinkWithPatternAndDomainPatternLarge-4                 69610             17136 ns/op            4958 B/op        151 allocs/op
PASS
ok      github.com/casbin/casbin/v2     76.995s
```

After:
```
goos: linux
goarch: amd64
pkg: github.com/casbin/casbin/v2
cpu: Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz
BenchmarkCachedRaw-4                                            76492984                15.80 ns/op            0 B/op          0 allocs/op
BenchmarkCachedBasicModel-4                                      7648494               154.5 ns/op           104 B/op          4 allocs/op
BenchmarkCachedRBACModel-4                                       7803536               154.3 ns/op           104 B/op          4 allocs/op
BenchmarkCachedRBACModelSmall-4                                  7738653               154.2 ns/op           104 B/op          4 allocs/op
BenchmarkCachedRBACModelMedium-4                                 7957747               150.8 ns/op           104 B/op          4 allocs/op
BenchmarkCachedRBACModelLarge-4                                  8025734               134.6 ns/op            96 B/op          3 allocs/op
BenchmarkCachedRBACModelWithResourceRoles-4                      7495996               156.4 ns/op           104 B/op          4 allocs/op
BenchmarkCachedRBACModelWithDomains-4                            7333226               163.9 ns/op           120 B/op          4 allocs/op
BenchmarkCachedABACModel-4                                        237619              5033 ns/op            2729 B/op         47 allocs/op
BenchmarkCachedKeyMatchModel-4                                   6933344               172.3 ns/op           152 B/op          4 allocs/op
BenchmarkCachedRBACModelWithDeny-4                               7765857               157.7 ns/op           104 B/op          4 allocs/op
BenchmarkCachedPriorityModel-4                                   7734268               157.1 ns/op           104 B/op          4 allocs/op
BenchmarkCachedRBACModelMediumParallel-4                        14936707                77.31 ns/op          105 B/op          4 allocs/op
BenchmarkHasPolicySmall-4                                        2330184               522.1 ns/op           182 B/op          8 allocs/op
BenchmarkHasPolicyMedium-4                                       2220385               549.2 ns/op           190 B/op          8 allocs/op
BenchmarkHasPolicyLarge-4                                        2119506               568.4 ns/op           197 B/op          9 allocs/op
BenchmarkAddPolicySmall-4                                        2256610               526.0 ns/op           184 B/op          8 allocs/op
BenchmarkAddPolicyMedium-4                                       1987681               591.9 ns/op           205 B/op          9 allocs/op
BenchmarkAddPolicyLarge-4                                        1000000              1088 ns/op             506 B/op         11 allocs/op
BenchmarkRemovePolicySmall-4                                     2235127               522.5 ns/op           198 B/op          9 allocs/op
BenchmarkRemovePolicyMedium-4                                    1924584               594.1 ns/op           209 B/op          9 allocs/op
BenchmarkRemovePolicyLarge-4                                      992588              1138 ns/op             322 B/op         15 allocs/op
BenchmarkRaw-4                                                  75003219                15.91 ns/op            0 B/op          0 allocs/op
BenchmarkBasicModel-4                                             125618              9525 ns/op            4945 B/op         90 allocs/op
BenchmarkRBACModel-4                                               92558             13187 ns/op            6332 B/op        116 allocs/op
BenchmarkRBACModelSmall-4                                          10000            112212 ns/op           53250 B/op       1000 allocs/op
BenchmarkRBACModelMedium-4                                          1113           1053600 ns/op          581872 B/op       9134 allocs/op
BenchmarkRBACModelLarge-4                                            100          10484457 ns/op         5177710 B/op      90380 allocs/op
BenchmarkRBACModelWithResourceRoles-4                              92574             12694 ns/op            7388 B/op        110 allocs/op
BenchmarkRBACModelWithDomains-4                                    79123             15507 ns/op            8374 B/op        138 allocs/op
BenchmarkABACModel-4                                              237844              5047 ns/op            2716 B/op         46 allocs/op
BenchmarkKeyMatchModel-4                                           82950             14393 ns/op            8153 B/op        122 allocs/op
BenchmarkRBACModelWithDeny-4                                       84972             14448 ns/op            6692 B/op        128 allocs/op
BenchmarkPriorityModel-4                                          109629             10966 ns/op            5849 B/op         99 allocs/op
BenchmarkRBACModelWithDomainPatternLarge-4                         35719             33435 ns/op           23269 B/op        274 allocs/op
BenchmarkRoleManagerSmall-4                                        28006             43793 ns/op           13540 B/op        498 allocs/op
BenchmarkRoleManagerMedium-4                                        2931            433114 ns/op          141913 B/op       5742 allocs/op
BenchmarkRoleManagerLarge-4                                          256           4487358 ns/op         1509928 B/op      59742 allocs/op
BenchmarkBuildRoleLinksWithPatternLarge-4                              1        1172280168 ns/op        689151360 B/op  10085486 allocs/op
BenchmarkBuildRoleLinksWithDomainPatternLarge-4                      160           7283604 ns/op         3786323 B/op      89050 allocs/op
BenchmarkBuildRoleLinksWithPatternAndDomainPatternLarge-4              1        1173982869 ns/op        690965336 B/op  10120419 allocs/op
BenchmarkHasLinkWithPatternLarge-4                                 73510             16953 ns/op            4883 B/op        153 allocs/op
BenchmarkHasLinkWithDomainPatternLarge-4                          965545              1267 ns/op             528 B/op         13 allocs/op
BenchmarkHasLinkWithPatternAndDomainPatternLarge-4                 64540             18024 ns/op            5312 B/op        162 allocs/op
PASS
ok      github.com/casbin/casbin/v2     76.942s
```

The major changes you can see are in the `BenchmarkRBACModel...` benchmarks, where we've gone:

<table>
<tr><th>Benchmark</th><th colspan=2>Before</th><th colspan=2>After</th></tr>
<tr><th>BenchmarkRBACModel-4</th><td>    6069 B/op</td><td>	     128 allocs/op</td><td>    6275 B/op</td><td>	     116 allocs/op</td></tr>
<tr><th>BenchmarkRBACModelSmall-4</th><td>   45013 B/op</td><td>	    1400 allocs/op</td><td>   53238 B/op</td><td>	    1000 allocs/op</td></tr>
<tr><th>BenchmarkRBACModelMedium-4</th><td>  458853 B/op</td><td>	   13127 allocs/op</td><td>  584824 B/op</td><td>	    9134 allocs/op</td></tr>
<tr><th>BenchmarkRBACModelLarge-4</th><td> 4399901 B/op</td><td>	  130263 allocs/op</td><td> 5180076 B/op</td><td>	   90378 allocs/op</td></tr>
<tr><th>BenchmarkRBACModelWithResourceRoles-4</th><td>    6778 B/op</td><td>	     118 allocs/op</td><td>    7346 B/op</td><td>	     110 allocs/op</td></tr>
<tr><th>BenchmarkRBACModelWithDomains-4</th><td>    8084 B/op</td><td>	     144 allocs/op</td><td>    8328 B/op</td><td>	     138 allocs/op</td></tr>
</table>
```

We have a little more memory being allocated in total, but with *significantly* fewer allocations.

I’m not sure how reflective this microbenchmark is of reality, because in one of our systems, a *significant* portion of the time is spent in `Sprintf`, and this will eliminate it entirely.